# docs/contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketHistoryPointResponse": {
        "description": "Single market history datapoint.",
        "properties": {
          "time": {
            "title": "Time",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "time",
          "value"
        ],
        "title": "MarketHistoryPointResponse",
        "type": "object"
      },
      "MarketIndexResponse": {
        "additionalProperties": true,
        "description": "Global index/ticker row for landing feed.",
        "properties": {
          "regularMarketChangePercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regularmarketchangepercent"
          },
          "regularMarketPrice": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regularmarketprice"
          },
          "shortName": {
            "title": "Shortname",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "regularMarketPrice",
          "regularMarketChangePercent",
          "shortName"
        ],
        "title": "MarketIndexResponse",
        "type": "object"
      },
...
      "MarketOverviewResponse": {
        "description": "Market overview payload.",
        "properties": {
          "bist": {
            "$ref": "#/components/schemas/MarketSeriesResponse"
          },
          "crypto": {
            "$ref": "#/components/schemas/MarketSeriesResponse"
          }
        },
        "required": [
          "bist",
          "crypto"
        ],
        "title": "MarketOverviewResponse",
        "type": "object"
      },
      "MarketSeriesResponse": {
        "additionalProperties": true,
        "description": "Market overview row.",
        "properties": {
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "currentValue": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentvalue"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/MarketHistoryPointResponse"
            },
            "title": "History",
            "type": "array"
          }
        },
        "required": [
          "currentValue",
          "change",
          "history"
        ],
        "title": "MarketSeriesResponse",
        "type": "object"
      },
      "MarketTickerResponse": {
        "additionalProperties": true,
        "description": "Ticker strip row.",
        "properties": {
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "changePercent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changepercent"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "name",
          "price",
          "change",
          "changePercent"
        ],
        "title": "MarketTickerResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "7/24 Finansal Sinyal ve Analiz REST API",
    "title": "Otonom Analiz API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/borsapy/market/indices": {
      "get": {
        "operationId": "quotes_borsapy_market_indices_get",
        "parameters": [
          {
            "in": "query",
            "name": "symbol",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 50,
              "minItems": 1,
              "title": "Symbol",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Quotes Borsapy Market Indices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Quotes",
        "tags": [
          "Private market data"
        ],
        "x-access-role": "admin"
      }
    },
...
    "/borsapy/market/overview": {
      "get": {
        "operationId": "overview_borsapy_market_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Overview Borsapy Market Overview Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Overview",
        "tags": [
          "Private market data"
        ],
        "x-access-role": "admin"
      }
    },
    "/borsapy/market/status": {
      "get": {
        "operationId": "status_borsapy_market_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Status Borsapy Market Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Status",
        "tags": [
          "Private market data"
        ],
        "x-access-role": "admin"
      }
    },
    "/borsapy/market/ticker": {
      "get": {
        "operationId": "ticker_borsapy_market_ticker_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Ticker Borsapy Market Ticker Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ticker",
        "tags": [
          "Private market data"
        ],
        "x-access-role": "admin"
      }
    },
...
    "/market/indices": {
      "get": {
        "description": "Landing sayfası için global endeks özet verisi döndürür.",
        "operationId": "get_market_indices_market_indices_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "borsapy",
              "pattern": "^(borsapy|legacy)$",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "description": "Global endeks sembolleri. Örnek: ?symbol=^GSPC&symbol=^NDX&symbol=XU100.IS",
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "default": [
                "^GSPC",
                "^NDX",
                "XU100.IS"
              ],
              "description": "Global endeks sembolleri. Örnek: ?symbol=^GSPC&symbol=^NDX&symbol=XU100.IS",
              "items": {
                "type": "string"
              },
              "title": "Symbol",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketIndexResponse"
                  },
                  "title": "Response Get Market Indices Market Indices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Market Indices",
        "tags": [
          "Market Data"
        ],
        "x-access-role": "admin"
      }
    },
...
    "/market/overview": {
      "get": {
        "operationId": "get_market_overview_market_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "borsapy",
              "pattern": "^(borsapy|legacy)$",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Market Overview",
        "tags": [
          "Market Data"
        ],
        "x-access-role": "admin"
      }
    },
    "/market/ticker": {
      "get": {
        "operationId": "get_market_ticker_market_ticker_get",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "borsapy",
              "pattern": "^(borsapy|legacy)$",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketTickerResponse"
                  },
                  "title": "Response Get Market Ticker Market Ticker Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Market Ticker",
        "tags": [
          "Market Data"
        ],
        "x-access-role": "admin"
      }
    },
...
  }
}